pyproject.toml
[tool.poetry]
name = "hpc-client"
version = "0.0.1"
description = "Python Client for High Performance Computing (HPC)"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flywheel-sdk = "^17.7.0"
jinja2 = "^3.1.3"
pyyaml = "^6.0.1"
pydantic = "^1.8.2"
typing-extensions = "^3.10.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"
flake8-colors = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
